requirements.txt
# Core dependencies
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
httpx>=0.24.1

# LangChain and LLM dependencies
langchain>=0.0.267
langchain-openai>=0.0.1
openai>=0.27.8
chromadb>=0.4.6

# Async support
aiohttp>=3.8.5
asyncio>=3.4.3

# Email processing
email-validator>=2.0.0
python-multipart>=0.0.6

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.1

# Logging and monitoring
loguru>=0.7.0

# Optional: For development
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.4.1